+++ ICSharpCode.Decompiler/CSharp/Transforms/IntroduceQueryExpressions.cs
@@ -25,6 +25,7 @@ using System.Linq;
 
 using ICSharpCode.Decompiler.CSharp.Syntax;
 using ICSharpCode.Decompiler.IL;
+using ICSharpCode.Decompiler.Semantics;
 
 namespace ICSharpCode.Decompiler.CSharp.Transforms
 {
